+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * fix: occasional startup crashes (https://github.com/zellij-org/zellij/pull/1706)
 * fix: gracefully handle SSH disconnects (https://github.com/zellij-org/zellij/pull/1710)
 * fix: handle osc params larger than 1024 bytes (https://github.com/zellij-org/zellij/pull/1711)
+* Terminal compatibility: implement faux scrolling when in alternate screen mode(https://github.com/zellij-org/zellij/pull/1678)
 
 ## [0.31.3] - 2022-08-18
 * HOTFIX: fix up-arrow regression